+++ nilearn/maskers/__init__.py
@@ -8,6 +8,7 @@ from .nifti_labels_masker import NiftiLabelsMasker
 from .nifti_maps_masker import NiftiMapsMasker
 from .nifti_masker import NiftiMasker
 from .nifti_spheres_masker import NiftiSpheresMasker
+from .surface_maskers import SurfaceLabelsMasker, SurfaceMasker
 
 __all__ = [
     "BaseMasker",
@@ -18,4 +19,6 @@ __all__ = [
     "NiftiMapsMasker",
     "NiftiMasker",
     "NiftiSpheresMasker",
+    "SurfaceMasker",
+    "SurfaceLabelsMasker",
 ]
